feat(api/products): add get products by category feature
BREAKING CHANGE:
Add new views collections to store products view count.
on every request to get a prodcut if the end user is not the admin then
the views will increase automatically. and after deleting a product the
views document for the product will automatically removed from the
database.

Controller for the /products/categories/:cat_name route will check the
cat_name and call required service then will send back the data to the
end user